implement retry for TopicWriter, extract Retrier into a separate file
