+++ ydb/src/retry.rs
@@ -1,0 +1,95 @@
+use num::pow;
+use std::time::Duration;
+use tracing::instrument;
+
+const DEFAULT_RETRY_TIMEOUT: Duration = Duration::from_secs(5);
+const INITIAL_RETRY_BACKOFF_MILLISECONDS: u64 = 1;
+
+#[derive(Debug)]
+pub(crate) struct RetryParams {
+    pub(crate) attempt: usize,
+    pub(crate) time_from_start: Duration,
+}
+
+// May be extend in feature
+#[derive(Default, Debug)]
+pub(crate) struct RetryDecision {
+    pub(crate) allow_retry: bool,
+    pub(crate) wait_timeout: Duration,
+}
+
+pub(crate) trait Retry: Send + Sync {
+    fn wait_duration(&self, params: RetryParams) -> RetryDecision;
+}
+
+#[derive(Debug)]
+pub(crate) struct TimeoutRetrier {
+    pub(crate) timeout: Duration,
+}
+
+impl Default for TimeoutRetrier {
+    fn default() -> Self {
+        Self {
+            timeout: DEFAULT_RETRY_TIMEOUT,
+        }
+    }
+}
+
+impl Retry for TimeoutRetrier {
+    #[instrument(ret)]
+    fn wait_duration(&self, params: RetryParams) -> RetryDecision {
+        let mut res = RetryDecision::default();
+        if params.time_from_start < self.timeout {
+            if params.attempt > 0 {
+                res.wait_timeout =
+                    Duration::from_millis(pow(INITIAL_RETRY_BACKOFF_MILLISECONDS, params.attempt));
+            }
+            res.allow_retry = (params.time_from_start + res.wait_timeout) < self.timeout;
+        };
+
+        res
+    }
+}
+
+pub(crate) struct NoRetrier {}
+
+impl Retry for NoRetrier {
+    #[instrument(skip_all)]
+    fn wait_duration(&self, _: RetryParams) -> RetryDecision {
+        RetryDecision {
+            allow_retry: false,
+            wait_timeout: Duration::default(),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{types_converters::try_vec_to_list_of_structs, ydb_struct};
+
+    #[test]
+    fn try_vec_empty() {
+        assert!(matches!(try_vec_to_list_of_structs(vec![]), Ok(None)));
+    }
+
+    #[test]
+    fn try_vec_same_structure() {
+        let values = vec![ydb_struct!("id" => 1), ydb_struct!("id" => 2)];
+        assert!(matches!(try_vec_to_list_of_structs(values), Ok(Some(_))));
+    }
+
+    #[test]
+    fn try_vec_different_structure() {
+        let values = vec![ydb_struct!("id" => 1), ydb_struct!("key" => 2)];
+        assert!(try_vec_to_list_of_structs(values).is_err());
+    }
+
+    #[test]
+    fn try_vec_non_struct() {
+        let values = vec![ydb_struct!("id" => 1), 42i64.into()];
+        assert!(try_vec_to_list_of_structs(values).is_err());
+
+        let values = vec![1i64.into()];
+        assert!(try_vec_to_list_of_structs(values).is_err());
+    }
+}
